Set new fields in VolumeGroupSnapshotContent objects
This patch improves the sidecar to set the fields introduced by the
VolumeGroupSnapshotContent v1beta2 API when dynamically provisioning a
VolumeGroupSnapshot.

This information is then used by the common snapshot controller to set
the status of the member VolumeSnapshotContent objects, allowing the
relative fields to be set even when the CSI driver does not implement
the ListSnapshots RPC call.